+++ packages/components/src/Tooltip/index.ts
@@ -25,3 +25,4 @@
  */
 
 export * from './Tooltip'
+export * from './useTooltip'